# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assistant API",
    "description": "AI assistant API — A2A protocol (Agent-to-Agent), chat, and workflow management.\n\n**Authentication**: protected endpoints require `Authorization: Bearer <token>`.\nThe token is set via `--auth-token` / `ASSISTANT_WEB_TOKEN` on the server.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.135"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local assistant server"
    }
  ],
  "paths": {
...
    "/api/orgs/{org_id}/spaces/{space_id}/interfaces": {
      "get": {
        "tags": [
          "interfaces"
        ],
        "summary": "`GET /api/orgs/{org_id}/spaces/{space_id}/interfaces` — list interface instances.",
        "operationId": "list_interfaces",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "description": "Space ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interface instances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterfaceInstanceResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_token": []
          },
          {
            "oauth2": [
              "spaces:read"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "interfaces"
        ],
        "summary": "`POST /api/orgs/{org_id}/spaces/{space_id}/interfaces` — create an interface instance.",
        "operationId": "create_interface",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "description": "Space ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterfaceInstanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceInstanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_token": []
          },
          {
            "oauth2": [
              "spaces:write"
            ]
          }
        ]
      }
    },
    "/api/orgs/{org_id}/spaces/{space_id}/interfaces/{id}": {
      "delete": {
        "tags": [
          "interfaces"
        ],
        "summary": "`DELETE /api/orgs/{org_id}/spaces/{space_id}/interfaces/{id}` — delete an interface instance.",
        "operationId": "delete_interface",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "description": "Space ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Interface instance ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_token": []
          },
          {
            "oauth2": [
              "spaces:write"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInterfaceInstanceRequest": {
        "type": "object",
        "required": [
          "interface_type"
        ],
        "properties": {
          "config": {
            "description": "JSON configuration (tokens, channels, etc.)."
          },
          "interface_type": {
            "type": "string",
            "description": "Interface type: `\"slack\"`, `\"matrix\"`, `\"mattermost\"`, `\"signal\"`, `\"nextcloud\"`."
          }
        }
      },
...
      "InterfaceInstanceResponse": {
        "type": "object",
        "required": [
          "id",
          "space_id",
          "interface_type",
          "config",
          "created_at"
        ],
        "properties": {
          "config": {},
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "interface_type": {
            "type": "string"
          },
          "space_id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token. Pass the token issued by the server via `Authorization: Bearer <token>`. Configure it with `--auth-token` / `ASSISTANT_WEB_TOKEN`."
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {
              "api_keys:read": "List API keys",
              "api_keys:write": "Create and revoke API keys",
              "conversations:read": "Read conversations",
              "conversations:write": "Create and send messages",
              "org:manage": "Full organization management",
              "personas:read": "Read personas",
              "personas:write": "Create and update personas",
              "skills:read": "Read skills",
              "skills:write": "Create and update skills",
              "spaces:read": "Read spaces",
              "spaces:write": "Create and update spaces",
              "users:read": "Read users",
              "users:write": "Create and update users"
            }
          },
          "clientCredentials": {
            "tokenUrl": "/oauth/token",
            "scopes": {
              "api_keys:read": "List API keys",
              "api_keys:write": "Create and revoke API keys",
              "conversations:read": "Read conversations",
              "conversations:write": "Create and send messages",
              "org:manage": "Full organization management",
              "personas:read": "Read personas",
              "personas:write": "Create and update personas",
              "skills:read": "Read skills",
              "skills:write": "Create and update skills",
...
  },
...
}